backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Werkzeug==2.0.1
pdf2image==1.16.0
google-cloud-vision==2.4.2
SQLAlchemy==1.4.23
psycopg2-binary==2.9.1
pandas==1.3.3
xlsxwriter==3.0.1
gunicorn==20.1.0